requirements.txt
flask
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy
pyjwt
flask-cors